Separate SwarmTest into files (partial classes)
